fix(flash_attn_prefill): widen mask indexing to int64_t (closes pp65536 i32 overflow)
Closes Wave 4 root cause identified by phase A research at
/tmp/cfa-cfa-20260427-adr005-wave4/phase-A-report.md.  The pp65536
first_decode_token=0 anomaly recorded in
project_long_prefill_parity_inverts.md was a deterministic i32
multiplication overflow in two D=256 prefill mask-indexing sites that
fire at seq_len_q >= 32768 with the rank-2 broadcast mask layout used
by forward_prefill_batched.rs.

Fix:

  flash_attn_prefill.metal:1487-1495
    Drop the narrowing `int(mask_params->M_strides[2])` cast on the
    str_x argument to MMAFrag_t::load_safe.  M_strides[2] is int64_t
    in AttnMaskParams (line 1054).  Passing it through preserves the
    full width into the template body's `(off_x + i) * str_x` product,
    which now widens to int64_t before the multiplication.

  flash_attn_prefill_blk.metal:181
    Cast both factors of `(qt * BQ) * M_stride` to int64_t.  The product
    can reach 4.29 G at qt=2047, BQ=32, M_stride=65536 — wraps i32 at
    qt >= 1024.

Mirrors the already-correct flash_attn_prefill_d512.metal:411-413
ulong-cast idiom (the D=512 path was unaffected because its prefill
kernel was ported from llama.cpp's wide-D path which already widens for
this exact reason).

Verification:

  Pre-patch  (commit 0d33301):
    pp65536_no_overflow_in_mask_indexing_last_row: 4/256 elements fail
    worst: idx=22 actual=-2.500000e-1 expected=-2.617188e-1
    abs_err=1.172e-2 rel_err=4.478e-2
    (budget: atol=5.000e-3 + rtol=2.000e-2)

  Post-patch (this commit):
    pp65536_no_overflow_in_mask_indexing_last_row: PASS
    max_abs=0.000e0 max_rel=0.000e0
    GPU output bf16-rounds bit-exact to CPU reference.

Full test suite:
  cargo test --release --test test_flash_attn_prefill: 49/49 pass
  cargo test --release --test test_flash_attn_vec:      5/5  pass
  cargo test --release --test test_flash_attn_vec_tq:   3/3  pass

  /opt/hf2q cargo check --release: clean compile
  (mlx-native is consumed via path patch in /opt/hf2q/.cargo/config.toml,
   so the shader change flows directly without a crate version bump).

Why no test ever caught this until now: existing flash_attn_prefill
tests covered seq_len in {8, 32, 128, 512, 2455} — the smallest seq_len
at which row_pos * kL exceeds i32::MAX is 32769.  The new pp65536 test
locks the regression to a deterministic synthetic GPU reproducer (no
model load required, ~9 GB peak alloc, 1.6s runtime).  Per
feedback_tests_on_gpu_not_cpu.md and feedback_no_vw_cheating.md.

Refs:
  - /tmp/cfa-cfa-20260427-adr005-wave4/phase-A-report.md (root-cause analysis)
  - flash_attn_prefill_d512.metal:411-413 (already-correct D=512 idiom)
  - mlx-native commit 0d33301 (failing-test-first commit; this fixes it)

+++ src/shaders/flash_attn_prefill_blk.metal
@@ -177,8 +177,16 @@ kernel void flash_attn_prefill_blk_bf16(
         // contribute identity values (bfloat16_t has no sentinel better
         // than what we start mmin/mmax at, so skipping the read is the
         // correct behaviour).
+        //
+        // Wave 4 Phase B: widen the row-offset multiplication to int64_t.
+        // With qt, BQ, M_stride all int the product `(qt * BQ) * M_stride`
+        // overflows i32 at qt >= 1024 (1024 * 32 * 65536 = 2^31), wrapping
+        // to a large negative pointer offset and reading garbage before
+        // the mask buffer's base.  Mirrors the already-correct
+        // flash_attn_prefill_d512.metal:411-413 ulong-cast idiom.  See
+        // /tmp/cfa-cfa-20260427-adr005-wave4/phase-A-report.md §2.5.2.
         device const bfloat16_t* mask_src =
-            mask + (qt * BQ) * M_stride + tile_k_start + tiisg;
+            mask + (int64_t)(qt * BQ) * (int64_t)M_stride + tile_k_start + tiisg;
 
         // Use f32 for reduction to avoid bf16 comparison subtleties — bf16
         // min/max are well-defined on Apple Silicon but f32 reductions are
